- Run yjs connection on a different path - Add zindex to cursor such that query editor has precedence
