+++ server/src/events/events.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { EventsGateway } from './events.gateway';
 import { YjsGateway } from './yjs.gateway';
 import { AuthModule } from 'src/modules/auth/auth.module';
 
 @Module({
   imports: [AuthModule],
-  providers: [YjsGateway],
+  providers: [EventsGateway, YjsGateway],
 })
 export class EventsModule {}
